testing_requirements.txt
httpx
bs4
PyYAML
crawl4ai
pydantic